increase JWT token lifetimes in settings
Increased the lifetimes of both access and refresh JWT tokens in the settings file. This change was necessary to improve the user experience by reducing the frequency of logins, providing a longer term authentication solution.Adjust JWT token lifetimes in HelloMeals

Extended the lifetimes of both access and refresh JWT tokens from seconds/minutes to days in the settings.py of HelloMeals. This change aims to enhance the user experience by decreasing the need for frequent logins, thereby offering a more efficient and long-term authentication approach.