Update color scheme for better contrast
The changes in this commit were made to update the color scheme of the web application. The color of text and certain elements were changed to improve visibility and contrast on darker backgrounds, enhancing the user interface design. In addition, unnecessary 'abc' text was removed and a margin-bottom was added for better layout design. Furthermore, the