Enhance recipe browsing experience
Made adjustments to the frontend to enhance the browsing experience for users. This includes updates to the recipeStore.ts to only fetch recipes from the server if the list of recipes is empty, unless a force update is requested. This will avoid unnecessary server requests and speed up the application.

Updated the recipe page layout to better present the available recipes, with a more inviting design, and added two new links to view all recipes and favorite recipes. Added related translation text for those links.

Fixed the navigation links in the layout to have a leading "/" which was causing issues with Nuxt.js routing.

Added two new pages for viewing all recipes and favorite recipes for the newly added links. These pages currently